requirements.txt
docker
requests==2.29.0
apache-airflow==2.5.0
apache-airflow-providers-google==10.0.0
psycopg2==2.9.6
pandas==2.0.2

